# Conflicts:
#	.idea/workspace.xml

+++ app/controller/historyEventController.php
@@ -42,28 +42,33 @@ class historyEventController
     }
     public function historyManagement($addError) {
 
-        if (($_SERVER["REQUEST_METHOD"] == 'POST') && isset($_POST['delete'])){
-            $this->deleteCardContent($_POST['id']);
-        } elseif (($_SERVER["REQUEST_METHOD"] == 'POST') && isset($_POST['deleteSchedule'])){
-            $this->deleteScheduleContent($_POST['tableId']);
+        if(isset($_SESSION['current_user']) && $_SESSION['current_user']->role == '3') {
+
+            if (($_SERVER["REQUEST_METHOD"] == 'POST') && isset($_POST['delete'])) {
+                $this->deleteCardContent($_POST['id']);
+            } elseif (($_SERVER["REQUEST_METHOD"] == 'POST') && isset($_POST['deleteSchedule'])) {
+                $this->deleteScheduleContent($_POST['tableId']);
+            }
+
+            if (($_SERVER["REQUEST_METHOD"] == 'POST') && isset($_POST['update'])) {
+                $this->updateCardContent($_POST['id'], $_POST['title'], $_POST['image'], $_POST['content']);
+            } else if (($_SERVER["REQUEST_METHOD"] == 'POST') && isset($_POST['updateSchedule'])) {
+                $this->updateScheduleContent($_POST['id'], $_POST['dateAndDay'], $_POST['time'], $_POST['language'], $_POST['ticketAmount']);
+            } else if (($_SERVER["REQUEST_METHOD"] == 'POST') && isset($_POST['updateMainContent'])) {
+                $this->updateMainContent($_POST['id'], $_POST['mainImageHeader'], $_POST['tourCardHeader'], $_POST['tourCardParagraph'], $_POST['tourCardButtonText']);
+            }
+
+            //used to get data from db to the view
+            $locations = $this->historyEventService->getAllHistoryCard();
+            $historyTourTimetable = $this->historyEventService->getHistoryTourTimetable();
+            $content = $this->historyEventService->getHistoryPageContent();
+
+            include __DIR__ . '/../view/header_history.php';
+            require __DIR__ . '/../view/history/historyAdmin/historyManagement.php';
         }
-
-        if (($_SERVER["REQUEST_METHOD"] == 'POST') && isset($_POST['update'])){
-            $this->updateCardContent($_POST['id'], $_POST['title'], $_POST['image'], $_POST['content']);
-        } else if (($_SERVER["REQUEST_METHOD"] == 'POST') && isset($_POST['updateSchedule'])) {
-            $this->updateScheduleContent($_POST['id'], $_POST['dateAndDay'], $_POST['time'], $_POST['language'], $_POST['ticketAmount']);
-        } else if (($_SERVER["REQUEST_METHOD"] == 'POST') && isset($_POST['updateMainContent'])) {
-            $this->updateMainContent($_POST['id'], $_POST['mainImageHeader'], $_POST['tourCardHeader'], $_POST['tourCardParagraph'], $_POST['tourCardButtonText']);
+        else {
+            $this->historyMainPage();
         }
-
-        //used to get data from db to the view
-        $locations = $this->historyEventService->getAllHistoryCard();
-        $historyTourTimetable = $this->historyEventService->getHistoryTourTimetable();
-        $content = $this->historyEventService->getHistoryPageContent();
-
-        include __DIR__ . '/../view/header_history.php';
-        require __DIR__ . '/../view/history/historyAdmin/historyManagement.php';
-
     }
 
     public function deleteScheduleContent($id)
